requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
groq==0.20.0
chromadb==0.4.22
sentence-transformers==2.5.1
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
python-multipart==0.0.9
datasets==2.17.1
tqdm==4.66.2

# Development dependencies
pytest==8.0.2
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 